fix: Support to serialize weights in AnimationCurve
